requirements.txt
Flask
Flask-Dance
python-dotenv
cryptography
